Decouple CpuFeatures from serializer state.
Traditionally, we cross compile a snapshot iff the serializer is enabled.
This will change in the future.

Changes:
 - CpuFeatures probing is done once per process, depending on whether we
   cross compile.
 - CpuFeatures are consolidated into the platform-independent assembler.h
   as much as possible.
 - FLAG_enable_<feature> will only be checked at probing time (already the
   case for ARM).
 - The serializer state is cached by the MacroAssembler.
 - PlatformFeatureScope is no longer necessary.
 - CPUFeature enum values no longer map to CPUID bit fields.

R=svenpanne@chromium.org

Review URL: https://codereview.chromium.org/285233010

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@21347 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/platform-posix.cc
@@ -52,7 +52,7 @@ namespace internal {
 static const pthread_t kNoThread = (pthread_t) 0;
 
 
-uint64_t OS::CpuFeaturesImpliedByPlatform() {
+unsigned OS::CpuFeaturesImpliedByPlatform() {
   return 0;  // Nothing special.
 }
